// src/app/services/openapi-client/model/updateOrderStatusDto.ts
/**
 * ShopErpApi | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO for updating order status
 */
export interface UpdateOrderStatusDto { 
    status: string;
    trackingNumber?: string | null;
    adminNotes?: string | null;
}